% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransAbess.R
\name{models}
\alias{models}
\title{Generate different target data and source data.}
\usage{
models(
  type = c("all", "source", "target"),
  beta = c(3, 1.5, 1, 0.5, 2, 2.5, 4, 3.5),
  p = 8,
  s = NULL,
  sig.strength = 1,
  cov.strength = 0.5,
  eps.strength = 0.3,
  h = 2,
  h.not.in = 6,
  K = 10,
  Ka = NULL,
  n.target = 200,
  n.source = rep(100, K)
)
}
\arguments{
\item{type}{different data to be generated. Can be either "all" or "source" or "target". 
\itemize{
\item "all": generate a list with a target data set of size \code{n.target} and \code{K} source data set of size \code{n.source}.
\item "source": generate a list with K source data set of size \code{n.source}.
\item "target": generate a list with a target data set of size \code{n.target}.
}}

\item{beta}{regression coefficient. Default = \code{c(3, 1.5, 1, 0.5, 2, 2.5, 4, 3.5)}.}

\item{p}{dimension of \code{beta}. Default is 8.}

\item{s}{sparse level of \code{beta}.}

\item{sig.strength}{signal strength in \code{beta}, i.e. the size of nonzero components.}

\item{cov.strength}{strength of elements in the covariance matrix.}

\item{eps.strength}{variance of the residuals.}

\item{h}{transferring level, i.e. the difference of the regression parameter between target data and transferring source data.}

\item{h.not.in}{to set the regression parameter of source data which can not be transferred.}

\item{K}{size of source data.}

\item{Ka}{size of transferring source data.}

\item{n.target}{sample size of target data.}

\item{n.source}{sample size of source data.}
}
\value{
a list containing target and source data.
}
\description{
Generate different target data and source data.
}
\examples{
\dontrun{
data <- models(type = 'all', K = 10, Ka = 5)
data <- models(type = 'all', sig.strength = 1, p = 500, s = 10, 
h = 5, h.not.in = 100, K = 10, Ka = 5)
data <- models(type = 'all', sig.strength = 1, beta = c(3, 1.5, 0, 0, 2, 0, 0, 0), K = 10, Ka = 5)
}
}
