\name{SA23204160-package}
\alias{SA23204160-package}
\alias{SA23204160}
\docType{package}
\title{
  Fit a transfer learning linear model based on abess algorithm
}
\description{
  R functions to implement TransAbess algorithm, with Rcpp functions used for faster matrices construction.
}
\author{
Liangchen He

Maintainer: Liangchen He <sayramoony@gmail.com>
}
\keyword{ package }
