% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransAbess.R
\name{TransAbess}
\alias{TransAbess}
\title{Fit a transfer learning linear model based on \code{abess} algorithm.}
\usage{
TransAbess(
  target,
  source = NULL,
  pred.target = NULL,
  transfer.source.id = "auto"
)
}
\arguments{
\item{target}{target data. Should be a list with elements x and y, where x indicates a predictor matrix with each row/column as a(n) observation/variable, and y indicates the response vector.}

\item{source}{source data. Should be a list with some sublists, where each of the sublist is a source data set, having elements x and y with the same meaning as in target data.}

\item{pred.target}{a new target data. Should be a list with elements x and y which we want to predict.}

\item{transfer.source.id}{transferable source indices. Can be either a subset of \code{{1, ..., length(source)}}, "all" or "auto". Default = \code{"auto"}.
\itemize{
\item a subset of \code{{1, ..., length(source)}}: only transfer sources with the specific indices.
\item "all": transfer all sources.
\item "auto": run transferable source detection algorithm to automatically detect which sources to transfer.
}}
}
\value{
a list named \code{TransAbess}, containing the parameter estimation \code{beta.hat} and the transferable source indices estimation \code{transfer.source.id} and the prediction error.
}
\description{
Fit a transfer learning linear model based on \code{abess} algorithm.
}
\examples{
\dontrun{
data <- models(type = 'all', sig.strength = 1, p = 500, s = 10, 
h = 5, h.not.in = 100, K = 10, Ka = 5)
transfer.source.id <- 1:5
fit.OracleTransABESS <- TransAbess(target = data$target, source = data$source, 
transfer.source.id = transfer.source.id)
}
}
